+++ socket_event_template/room_manager.h
@@ -2,24 +2,24 @@
 #define ROOM_MANAGER_H_
 
 #include "chat_room.h"
+#include "../common/auto_lock.h"
 #include <iostream>
 #include <map>
 #include <pthread.h>
-//lock
 
 class RoomManager
 {
 private:
-	std::map<long,ChatRoom*> room_list_;
+	std::map<int,ChatRoom*> room_list_;
 	pthread_mutex_t list_mutex_;
 
 public:
 	RoomManager();
 	~RoomManager();
 
-	bool AddRoom(long room_id,int server_fd);
-	ChatRoom * GetChatRoom(long room_id);
-	bool DelRoom(long room_id);
+	bool AddRoom(int room_id,int server_fd);
+	ChatRoom * GetChatRoom(int room_id);
+	bool DelRoom(int room_id);
 };
 
 #endif
